+++ lib/grape/exceptions/invalid_accept_header.rb
@@ -1,0 +1,10 @@
+# encoding: utf-8
+module Grape
+  module Exceptions
+    class InvalidAcceptHeader < Base
+      def initialize(message, headers)
+        super(message: compose_message('invalid_accept_header', message: message), status: 406, headers: headers)
+      end
+    end
+  end
+end